conda_requirements.txt
# ############################################################################
# conda_requirements.txt
# ======================
# Author : Sepand KASHANI [sepand.kashani@epfl.ch]
# ############################################################################

# This file contains a superset of Acoustic Cameras's dependencies listed in
# `requirements.txt` that can be installed through conda packages.
# These packages are the preferred way to install dependencies if available.
#
# When modifying dependencies, don't forget to update both `requirements.txt`
# and `conda_requirements.txt`.

python == 3.6.*
astropy == 3.0.*
ipython == 7.2.*
matplotlib == 3.0.*
numexpr == 2.6.*
numpy == 1.15.*
pandas == 0.23.*
pyproj == 1.9.*
pyunlocbox == 0.5.*
redis-py == 3.2.*
scikit-image == 0.14.*
scipy == 1.1.*
tqdm == 4.28.*
